+++ pages/api/forms/pit/[id].ts
@@ -15,10 +15,10 @@ export default new RouteHandler<'api', WAuth>()
 		return res.status(200).json({ form, images });
 	})
 	.delete(async (req, res) => {
-		if (!req.user.administrator || !req.user)
-			return res.status(401).json({ message: 'You are not authorized to update users.' });
+		if (!req.user.administrator || !req.user || req.user.banned)
+			return res.status(401).json({ message: 'You are not authorized to delete pit forms.' });
 		const id = String(req.query.id);
-		PitForm.findByIdAndDelete(id);
-		PitImage.deleteMany({ formId: id });
+		await PitForm.findByIdAndDelete(id);
+		await PitImage.deleteMany({ formId: id });
 		return res.status(200).json({ message: 'Successfully deleted.' });
 	});
